services/ml-service/requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Sudoku Ultra — ML Service Dependencies
# ──────────────────────────────────────────────────────────────────────────────

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# ML / AI
torch==2.2.0
torchvision==0.17.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2

# Explainability + Tuning
shap==0.45.0
optuna==3.6.0

# MLflow
mlflow==2.14.0

# CV
opencv-python-headless==4.10.0.84

# Model Export
onnx==1.16.0
onnxruntime==1.18.0
skl2onnx==1.17.0
# TFLite export (optional, heavy): pip install onnx2tf tensorflow

# Analytics
duckdb==0.10.3

# Kafka consumer
kafka-python==2.0.2

# Monitoring
evidently==0.4.27

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9

# NLP / Toxicity classifier
transformers==4.41.0
datasets==2.19.0
accelerate==0.30.0     # HuggingFace Trainer accelerator backend

# RL Bot
stable-baselines3==2.3.0
gymnasium==0.29.1
sb3-contrib==2.3.0   # MaskablePPO for valid-action masking

# RAG / LLM
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-huggingface==0.0.6
qdrant-client==1.10.1
sentence-transformers==3.1.1

# MinIO / S3 artifact store
boto3==1.34.144

# Observability
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
sentry-sdk[fastapi]==2.8.0

# Analytics / rate limiting
slowapi==0.1.9
great-expectations==0.18.14

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx  # also used as async test client
